+++ src/app/components/shared/autocomplete/AutocompleteOption.model.ts
@@ -1,4 +1,5 @@
 export interface AutocompleteOption {
   label: string;
   value: number | string;
+  description?: string;
 }